requirements.txt
sentence_transformers
chromadb
openai
requests
python-dotenv
tqdm
aiohttp
pypdf
python-docx
pandas
openpyxl
alembic==1.18.1
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
boto3==1.42.37
botocore==1.42.37
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jmespath==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.7
langsmith==0.6.4
Mako==1.3.10
MarkupSafe==3.0.3
memu-py==1.2.0
numpy==2.4.1
openai==2.15.0
orjson==3.11.5
packaging==25.0
pendulum==3.1.0
pgvector==0.4.2
pip==25.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
sqlmodel==0.0.31
tenacity==9.1.2
tencentcloud-sdk-python==3.1.38
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==1.26.20
uuid_utils==0.14.0
websockets==16.0
zstandard==0.25.0
